Updates for creation of PDF file with PDFBox
