+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/exceptions/VerifyTokenServiceException.java
@@ -1,9 +1,15 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.service.exceptions;
 
+import java.io.Serial;
+
 /**
  * Is thrown when verify token service fails to verify the JWT.
  */
 public class VerifyTokenServiceException extends RuntimeException {
+
+    @Serial
+    private static final long serialVersionUID = 9045863543269746292L;
+
     /**
      * Creates a {@link VerifyTokenServiceException} with only a message.
      * @param msg a message explaining why this exception is thrown
